+++ src/core/debug/examples.js
@@ -158,12 +158,27 @@ const exampleScreens = {
 			next: "debug",
 		},
 	},
-
 	"dom-text": {
 		scene: DomText,
 		title: "DOM Text",
 		routes: {
 			back: "debug",
+			next: "debug",
+		},
+	},
+	basisu: {
+		scene: Home,
+		title: "Basis Universal",
+		routes: {
+			debug: "debug",
+			next: "debug",
+		},
+	},
+	"basisu-quality": {
+		scene: Home,
+		title: "Basis Universal Quality",
+		routes: {
+			debug: "debug",
 			next: "debug",
 		},
 	},