+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 | Version       | Description                                                                                                                                           |
 |---------------|-------------------------------------------------------------------------------------------------------------------------------------------------------|
+|               | Fix crash if no levels collection has been defined.                                                                                                   |
 |               | Make loadbarPosY optional.                                                                                                                            |
 |               | Default to GL Mode for Kindles.                                                                                                                       |
 |               | Add Basis Universal Loader plugin for Phaser.                                                                                                         |