+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 | Version       | Description                                                                                                                                           |
 |---------------|-------------------------------------------------------------------------------------------------------------------------------------------------------|
 |               | Added `addOuterStroke` method to dom text - automatically adds pseudo element hacks to simulate outer stroke.                                         |
+|               | Updated Phaser to 3.70.                                                                                                                               |
 |               | Update Crel lib and move to npm dependency now it is available as an ES6 module.                                                                      |
 |               | Fix slow start with `npm run start`.                                                                                                                  |
 |               | Fix crash if no levels collection has been defined.                                                                                                   |